requirements.txt
Flask==2.3.3
flask-socketio==5.3.4
flask-cors==4.0.0
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
pyaudio==0.2.13
python-socketio==5.9.0
eventlet==0.33.3
protobuf==4.24.3